replace icon images, disable autocomplete on forms

+++ my-app/src/components/DeleteCardButton.js
@@ -2,45 +2,35 @@ import React from "react";
 import { Button } from 'semantic-ui-react'
 import { useNavigate } from "react-router-dom";
 
-function DeleteCardButton( { id }) {
+function DeleteCardButton( { id, deleteLog }) {
     const button = <Button secondary onClick={()=>handleClick(id)} style={{background: "#BF0000"}}>Delete Log</Button>
-
+    const navigate = useNavigate();
 
     console.log(id)
 
     function handleClick(id){
+       const confirm= window.confirm("This log will be deleted.");
 
-            fetch(`http://localhost:3000/reflections/${id}`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json"
-                }
-            })
-            .then(res => res.json())
-            .then(data => console.log(data))        
-
-    //    const confirm= window.confirm("This log will be deleted.");
-
-    //    if (confirm) {
-    //         useEffect(() => {
-    //             fetch(`http://localhost:3000/reflections/${id}`, {
-    //                 method: "DELETE",
-    //                 headers: {
-    //                     "Content-Type": "application/json"
-    //                 }
-    //             })
-    //             .then(res => res.json())
-    //             .then(data => console.log(data))        
-    //         }, [id])
-    //    } else {
-    //         console.log("Log not deleted")
-    //    }
-    //    console.log(confirm);
+       if (confirm) {
+                fetch(`http://localhost:3000/reflections/${id}`, {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json"
+                    }
+                })
+                .then(res => {
+                    if (res.ok) {
+                        deleteLog(id);
+                        navigate("/logs")
+                    }
+                })    
+       } else {
+            console.log("Log not deleted")
+       }
+       console.log(confirm);
     }
 
 
-
-    
     
     return(
         <div>
